Several changes:
1: added a field to FACETS in config.py in catalog called "facetlocation" that allows people to specify where they want their facets to be show on the template. It looks like this:
FACETS = [    { 'name' : _('Dewey Call Number Range 100s'), 'code' : 'callnumlayerone', 'sortbycount' : 'false' , 'facetlocation' : 'sidebar'},

Currently "sidebar" and topbar" are specified in the template search.html (sidebar is the default), but the template can be changed to display however you want. Changing one of the facets from sidebar to topbar will put it on top. This is a start to allowing users to display dewey/LC call number ranges like NCSU. I haven't worked out how to make the results distrubute themselves across the screen in columns instead of as a list. In the index.html, I modified the CSS section to include a topbarfacets section to display the results.


2. To allow users the ability to "browse into" the LC or Dewey call number sections, I added two fields to the solr index. 'callnumlayerone' and 'callnumlayertwo'. I created a deweycallnumlayer1 and 2 processor to place items in dewey ranges. An LCcallnumlayer1 and 2 can be made, but I didn't make them because we don't use LC. The views.py then has some logic in it so that there is a one layer drilldown available. For example, if someone performs an initial search, the facets of the 100's dewey will be displayed and the 10's wouldn't (callnumlayer1 and 2 respectively). In the results, if they clicked on the 100's facet "100-199 - Philosophy and psychology", the next resulting screen would NOT display the 100's, but display the 10's. It currently supports two layer drilling and all that needs to be done for LC is to create a processor pull out the information from the MARC records and place them into 'callnumlayerone' and 'callnumlayertwo'. The logic in the views.py basically says that if callnumlayer1 is a limit, then display callnumlayertwo and delete callnumlayerone facets from the results. If callnumlayer1 is not a limit, delete callnumlayer2 results from the facets. It's slightly more complicated, but that's the gist of it. It's not perfect, but it's a start.
NOTE: IN THIS UPDATE THIS IS TURNED OFF BY DEFAULT. TO ACTIVATE THIS, UNCOMMENT CALLNUMLAYERONE AND TWO FROM THE INDEXES.PROPERTIES FILE AND RE-INDEX.